Allow users to select score region on a map (#312)
* Users can draw a polygon on the map to delineate the score region. Once drawn, they can modify the bounds, or delete the polygon.
* Put help icon next to drawing controls in standard map, as opposed to next to search box
* Move some helper functions that were only used in tests into prod.
* Factored out a common map creation class.
* Split manage_disaster_test.js into three test classes, since they all have different setup requirements:
  - The tests that called functions in create_score_asset.js went into create_score_asset_test.js. They were basically unchanged.
  - The tests that related to adding/deleting disasters went into manage_disaster_add_delete_test.js. They were also basically unchanged, except that I cy.wrap()ped the Promises again, because I realized it wasn't safe to not return them and not wrap them, and returning them made Cypress complain a bit.
  - The tests that checked score parameter validation went into manage_disaster_set_score_parameters_test.js. The bodies of these tests didn't change too much (map bounds validation is different), but the setup changed significantly. We're now only doing setup once per the whole test file, not once per test case, because the map is too brittle to re-initialize on a per-test-case basis. As compensation, we're doing a much more realistic setup, calling into the prod `enableWhenFirestoreReady`.
* Give up on a common document.* fakeout: I tried here, and it's too brittle (different prod libraries that we're using get confused in different ways). Better to let each test just fake out what it needs.

There's a fair amount of complexity introduced because the map is sensitive to not being visible. Since we can make the map invisible by switching the damage asset value, and that can happen very quickly on page load, we need a two-step process to prevent the map from complaining. I think it's still possible for the user to cause problems by unselecting a damage asset and then re-selecting it very quickly, but that is not likely in practice.

TODO: change documentation for map bounds selection once PR is in.

Most of #283.
