+++ docs/create_map.js
@@ -1,5 +1,5 @@
+import {createBasicMap} from './basic_map.js';
 import {POLYGON_HELP_URL} from './help.js';
-import mapStyles from './map_styles.js';
 import {geoPointToLatLng} from './map_util.js';
 import {setUpPolygonDrawing} from './polygon_draw.js';
 
@@ -20,13 +20,7 @@ const placeIconParams = {
  * @return {google.maps.Map}
  */
 function createMap(firebasePromise) {
-  // Create the base Google Map. Takes ~7 ms to execute this step.
-  // Temporarily center on the center of the continental 48 states.
-  // In practice, the firebase promise finishes so fast we don't actually
-  // see this happen.
-  const map = new google.maps.Map(
-      $('.map').get(0),
-      {center: {lat: 39.8283, lng: -98.5795}, styles: mapStyles});
+  const {map, searchBox} = createBasicMap(document.getElementById('map'));
 
   firebasePromise.then((doc) => {
     const mapBounds = doc.data()['map-bounds'];
@@ -48,29 +42,19 @@ function createMap(firebasePromise) {
   helpIcon.setAttribute('aria-hidden', 'true');
   helpLink.appendChild(helpIcon);
   helpContainer.appendChild(helpLink);
-  map.controls[google.maps.ControlPosition.TOP_LEFT].push(helpContainer);
+  map.controls[google.maps.ControlPosition.TOP_LEFT].insertAt(0, helpContainer);
 
   // Search box code roughly taken from
   // https://developers.google.com/maps/documentation/javascript/examples/places-searchbox.
 
-  // Create the search box.
-  // Normally we would just steal this element from the html, but the map does
-  // weird grabby things with it, which don't seem worth working around.
-  const input = document.createElement('INPUT');
-  input.setAttribute('type', 'text');
-  input.setAttribute('placeholder', 'Search');
-  const searchBox = new google.maps.places.SearchBox(input);
-  map.controls[google.maps.ControlPosition.TOP_LEFT].push(input);
-
   // Bias the SearchBox results towards current map's viewport.
-  map.addListener('bounds_changed', function() {
-    searchBox.setBounds(map.getBounds());
-  });
+  map.addListener('bounds_changed', () => searchBox.setBounds(map.getBounds()));
 
   let markers = [];
   // Listen for the event fired when the user selects a prediction and retrieve
-  // more details for that place.
-  searchBox.addListener('places_changed', function() {
+  // more details for that place. This will fire along with the default handler
+  // added in createBasicMap.
+  searchBox.addListener('places_changed', () => {
     const places = searchBox.getPlaces();
 
     if (!places.length) {
@@ -82,30 +66,18 @@ function createMap(firebasePromise) {
     markers = [];
 
     // For each place, get the icon, name and location.
-    const bounds = new google.maps.LatLngBounds();
-    places.forEach(function(place) {
-      if (!place.geometry) {
-        console.log('Returned place contains no geometry');
-        return;
-      }
-
-      // Create a marker for each place.
-      markers.push(new google.maps.Marker({
-        map: map,
-        // Take the default params and add in the place-specific icon.
-        icon: Object.assign({url: place.icon}, placeIconParams),
-        title: place.name,
-        position: place.geometry.location,
-      }));
-
-      if (place.geometry.viewport) {
-        // Only geocodes have viewport.
-        bounds.union(place.geometry.viewport);
-      } else {
-        bounds.extend(place.geometry.location);
+    places.forEach((place) => {
+      if (place.geometry && place.geometry.location) {
+        // Create a marker for each place.
+        markers.push(new google.maps.Marker({
+          map: map,
+          // Take the default params and add in the place-specific icon.
+          icon: Object.assign({url: place.icon}, placeIconParams),
+          title: place.name,
+          position: place.geometry.location,
+        }));
       }
     });
-    map.fitBounds(bounds);
   });
   return map;
 }
