+++ docs/highlight_features.js
@@ -1,3 +1,4 @@
+import {applyMinimumBounds} from './basic_map.js';
 import {geoidTag} from './property_names.js';
 
 export {currentFeatures, CurrentFeatureValue, highlightFeatures};
@@ -84,19 +85,6 @@ function highlightFeatures(features, map, zoom = false) {
           (feature) => feature.getGeometry().forEachLatLng(
               (latlng) => bounds.extend(latlng)));
     }
-    // Make sure we're sufficiently zoomed out for reasonable map context.
-    const extendPoint1 = new google.maps.LatLng(
-        Math.max(
-            bounds.getNorthEast().lat(), bounds.getSouthWest().lat() + 0.1),
-        Math.max(
-            bounds.getNorthEast().lng(), bounds.getSouthWest().lng() + 0.1));
-    const extendPoint2 = new google.maps.LatLng(
-        Math.min(
-            bounds.getSouthWest().lat(), bounds.getNorthEast().lat() - 0.1),
-        Math.min(
-            bounds.getSouthWest().lng(), bounds.getNorthEast().lng() - 0.1));
-    bounds.extend(extendPoint1);
-    bounds.extend(extendPoint2);
-    map.fitBounds(bounds);
+    applyMinimumBounds(bounds, map);
   }
 }
