+++ docs/import/manage_disaster_startup.js
@@ -4,7 +4,7 @@ import {getDisastersData} from '../firestore_document.js';
 import {loadNavbarWithPicker} from '../navbar.js';
 import TaskAccumulator from '../task_accumulator.js';
 
-import {enableWhenReady, onSetDisaster, setUpScoreSelectorTable, toggleState} from './manage_disaster.js';
+import {enableWhenReady, onSetDisaster, setUpScoreBoundsMap, setUpScoreSelectorTable, toggleState} from './manage_disaster.js';
 
 // Two tasks: EE and page load. Firebase is taken care of in the promise.
 const taskAccumulator =
@@ -20,5 +20,6 @@ $(() => {
   $('#create-new-disaster').on('click', () => toggleState(false));
   $('#cancel-new-disaster').on('click', () => toggleState(true));
   setUpScoreSelectorTable();
+  setUpScoreBoundsMap(document.getElementById('score-bounds-map'));
   taskAccumulator.taskCompleted();
 });
