+++ docs/import/score_path_lib.js
@@ -1,0 +1,5 @@
+export {scoreCoordinatesAttribute};
+
+// TODO(janakr): add functions for create_score_asset use, and move property
+//  paths into this common library.
+const scoreCoordinatesAttribute = 'score_bounds_coordinates';
