+++ docs/index.html
@@ -67,7 +67,7 @@
     </div>
     <div class="main-content">
       <div class="map-container" id="mapContainer">
-        <div class="map"></div>
+        <div class="map" id="map"></div>
         <div class="loader" id="mapContainer-loader">
           <div class="ellipsis"><div></div><div></div><div></div><div></div></div>
         </div>
